исправление багов
