+++ project/src/components/film-list/film-list.tsx
@@ -1,12 +1,15 @@
-import FilmCard from '../card/card';
-import GenresFilter from '../genres-filter/genres-filter';
-import { useAppSelector } from '../../hooks';
-import ShowMoreButton from '../show-more-button/show-more-button';
-import { getCardCount, getFilteredFilms } from '../../store/main-data/selectors';
+import {useAppSelector} from '../../hooks';
 
-function Catalog(): JSX.Element {
+import GenresFilter from '../genres-filter/genres-filter';
+import FilmCard from '../film-card/film-card';
+import ShowMoreButton from '../show-more-button/show-more-button';
+import {getCardCount, getFilteredFilms} from '../../store/main-data/selectors';
+
+function FilmList(): JSX.Element {
+
   const films = useAppSelector(getFilteredFilms);
   const cardCount = useAppSelector(getCardCount);
+
   return (
     <section className="catalog">
       <h2 className="catalog__title visually-hidden">Catalog</h2>
@@ -18,9 +21,11 @@ function Catalog(): JSX.Element {
           <FilmCard
             key={film.id}
             id={film.id}
+
             title={film.name}
             image={film.previewImage}
             previewVideo={film.previewVideoLink}
+
           />))}
       </div>
 
@@ -29,4 +34,4 @@ function Catalog(): JSX.Element {
   );
 }
 
-export default Catalog;
+export default FilmList;
