+++ project/src/components/film-tab-item/film-tab-item.tsx
@@ -1,5 +1,5 @@
-import { useAppDispatch } from '../../hooks';
-import { changeFilmTab } from '../../store/film-data/film-data';
+import {useAppDispatch} from '../../hooks';
+import {changeFilmTab} from '../../store/film-data/film-data';
 
 type tabProps = {
   currentTab: string,
