+++ project/src/components/film-tabs/film-tabs.tsx
@@ -1,11 +1,11 @@
-import { FilmPageTabs } from '../../const';
+import {FilmPageTabs} from '../../const';
 import FilmTabItem from '../film-tab-item/film-tab-item';
 
 type tabsProps = {
   currentTab: string;
 }
 
-function FilmTabs({ currentTab }: tabsProps): JSX.Element {
+function FilmTabs ({currentTab}: tabsProps): JSX.Element {
   return (
     <nav className="film-nav film-card__nav">
       <ul className="film-nav__list">
