+++ project/src/components/history-route/history-route.tsx
@@ -8,11 +8,12 @@ export interface HistoryRouterProps {
   children?: React.ReactNode
 }
 
-function HistoryRouter({
-  basename,
-  children,
-  history,
-}: HistoryRouterProps) {
+function HistoryRouter(
+  {
+    basename,
+    children,
+    history,
+  }: HistoryRouterProps) {
   const [state, setState] = useState({
     action: history.action,
     location: history.location,
