+++ project/src/components/review-list/review-list.tsx
@@ -29,6 +29,7 @@ function ReviewList({reviews}: ReviewsProps): JSX.Element {
           />))}
       </div>
     </div>
+
   );
 }
