+++ project/src/components/user-block/user-block.tsx
@@ -1,13 +1,14 @@
-import { Link } from 'react-router-dom';
-import { logoutAction } from '../../store/api-actions';
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { AppRoute, AuthorizationStatus } from '../../const';
-import { getAuthorizationStatus, getAvatarUrl } from '../../store/user-process/selectors';
+import {Link} from 'react-router-dom';
+import {logoutAction} from '../../store/api-actions';
+import {useAppDispatch, useAppSelector} from '../../hooks';
+import {AppRoute, AuthorizationStatus} from '../../const';
+import {getAuthorizationStatus, getAvatarUrl} from '../../store/user-process/selectors';
 
 function UserBlock(): JSX.Element {
   const dispatch = useAppDispatch();
   const avatar = useAppSelector(getAvatarUrl);
   const authStatus = useAppSelector(getAuthorizationStatus);
+
   if (authStatus !== AuthorizationStatus.Auth) {
     return (
       <ul className="user-block">
@@ -22,12 +23,13 @@ function UserBlock(): JSX.Element {
       </ul>
     );
   }
+
   return (
     <ul className="user-block">
       <li className="user-block__item">
         <div className="user-block__avatar">
           <Link to={AppRoute.MyList}>
-            <img src={avatar || ''} alt="User avatar" width="63" height="63" />
+            <img src={avatar || ''} alt="User avatar" width="63" height="63"/>
           </Link>
         </div>
       </li>
