+++ project/src/index.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import {Provider} from 'react-redux';
+
+import {store} from './store';
+
 import App from './components/app/app';
-import { Provider } from 'react-redux';
-import { store } from './store';
 import ErrorMessage from './components/error-message/error-message';
+
 import {checkAuthAction, fetchFilmsAction, fetchPromoAction} from './store/api-actions';
 
 store.dispatch(fetchFilmsAction());
