+++ project/src/pages/main-page/main-page.tsx
@@ -1,12 +1,11 @@
-import { useEffect } from 'react';
-import Catalog from '../../components/catalog/catalog';
 import Logo from '../../components/logo/logo';
 import PromoCard from '../../components/promo-card/promo-card';
-import { AuthorizationStatus } from '../../const';
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { fetchFavoriteFilmsAction } from '../../store/api-actions';
-import { getAuthorizationStatus } from '../../store/user-process/selectors';
-
+import FilmList from '../../components/film-list/film-list';
+import {useAppDispatch, useAppSelector} from '../../hooks';
+import {useEffect} from 'react';
+import {getAuthorizationStatus} from '../../store/user-process/selectors';
+import {AuthorizationStatus} from '../../const';
+import {fetchFavoriteFilmsAction} from '../../store/api-actions';
 
 function MainPage(): JSX.Element {
   const dispatch = useAppDispatch();
@@ -17,14 +16,16 @@ function MainPage(): JSX.Element {
       dispatch(fetchFavoriteFilmsAction());
     }
   }, [authStatus, dispatch]);
+
   return (
     <>
       <PromoCard />
-      <div className="page-content">
-        <Catalog />
-        <footer className="page-footer">
-          <Logo lightMode/>
 
+      <div className="page-content">
+        <FilmList />
+
+        <footer className="page-footer">
+          <Logo isLightMode />
           <div className="copyright">
             <p>© 2019 What to watch Ltd.</p>
           </div>
@@ -33,4 +34,5 @@ function MainPage(): JSX.Element {
     </>
   );
 }
+
 export default MainPage;
