+++ project/src/services/process-error-handle.ts
@@ -1,6 +1,7 @@
-import { store } from '../store';
-import { setError } from '../store/app-process/app-process';
-import { TIMEOUT_SHOW_ERROR } from '../const';
+import {store} from '../store';
+
+import {setError} from '../store/app-process/app-process';
+import {TIMEOUT_SHOW_ERROR} from '../const';
 
 export const processErrorHandle = (message: string): void => {
   store.dispatch(setError(message));
