+++ project/src/store/action.ts
@@ -1,3 +1,3 @@
-import { createAction } from '@reduxjs/toolkit';
+import {createAction} from '@reduxjs/toolkit';
 
 export const redirectToRoute = createAction<string>('app/redirectToRoute');
