+++ project/src/store/index.ts
@@ -1,6 +1,6 @@
 import {configureStore} from '@reduxjs/toolkit';
-import { createAPI } from '../services/api';
-import { redirect } from './middleware/redirect';
+import {createAPI} from '../services/api';
+import {redirect} from './middlewares/redirect';
 import {rootReducer} from './root-reducer';
 
 export const api = createAPI();
