+++ project/src/types/auth-data.ts
@@ -1,4 +1,4 @@
 export type AuthData = {
-    email: string,
-    password: string,
-  };
+  email: string,
+  password: string,
+};
