+++ project/src/types/user-data.ts
@@ -1,7 +1,7 @@
 export type UserData = {
-    avatarUrl: string,
-    email: string,
-    id: number,
-    name: string,
-    token: string,
-  };
+  avatarUrl: string,
+  email: string,
+  id: number,
+  name: string,
+  token: string,
+};
