+++ project/src/utils/get-all-genres.ts
